+++ docs/CHANGELOG.md
@@ -12,12 +12,14 @@
 - fixed global RegExp in StringUtils.createReplacementsFromConfigArray
 - itemsjs: added spatial-search
 - geoutils: added distance-calculation
+- introduced fully featured markdown-support with extensions
 
 ### new features
 - introduced fully featured pdoc-services derived from cdoc
 - moved track-colors to commons
 - itemsjs: added spatial-search
 - geoutils: added distance-calculation
+- introduced fully featured markdown-support with extensions
 
 ### improvements
 - improved SqlAdapter with remapFulltextFilter